Started conversion to Spring Framework for commands
Reworked SetEvent and Invoke strategies in the Command class
Updated how slash command interactions are triggered. Now creates a new instance of the appropriate command for every event fire.
Created Command factory class which is the new way in which commands are supposed to be created
Command class now has logging delegate functions and a new way of interaction with channels.
Added reportTime() function in the Bot file for profiling